MedadWebsite
Medad is a website designed to connect readers, writers, and cultural organizations by gathering literary clubs and cultural events in one unified platform. We have developed the user interface for the website to ensure a smooth and intuitive experience. The platform allows users to explore and register for cultural events, join literary clubs, and manage their profiles easily, while offering a centralized space for discovering and showcasing cultural activities across the Kingdom.